+++ lib/src/request.dart
@@ -1,6 +1,8 @@
 import 'channel_owner.dart';
 import 'generated/channels.dart';
 
+/// Whenever the page sends a request for a network resource the following sequence of events are emitted by Page:
+/// - `page.on('request')` emitted when the request is issued by the page.
 class Request extends RequestBase {
   Request(
     super.connection,
@@ -10,9 +12,16 @@ class Request extends RequestBase {
     super.parent,
   ]);
 
+  /// URL of the request.
   String get url => initializer['url'] as String;
+
+  /// Resource type of the request.
   String get resourceType => initializer['resourceType'] as String;
+
+  /// HTTP method of the request (e.g. GET, POST).
   String get method => initializer['method'] as String;
+
+  /// Request's post body, if any.
   String? get postData => initializer['postData'] as String?;
   Map<String, dynamic> get headers =>
       (initializer['headers'] as List?)?.fold<Map<String, dynamic>>({}, (
@@ -27,6 +36,7 @@ class Request extends RequestBase {
   bool get isNavigationRequest =>
       initializer['isNavigationRequest'] as bool? ?? false;
 
+  /// Returns the matching Response object, or null if the response was not received due to error.
   Future<dynamic> response() async {
     final result = await channel_response();
     final resp = result.response;